fix(consume): allow absolute paths for `--evm-bin` (#1052)
* fix(consume): enable --evm-bin with absolute paths

If the value passed to --evm-bin was an abolute path, it got ignored by pytest. Perhaps, it got interpreted as a test paths? This workaround avoids the problem.

* chore(consume): better default value for --evm-bin with consume direct

The default value of the TransitionTool (`ethereum-spec-evm-resolver`) doesn't support state- or blocktest consumption; use the Geth `evm` binary in PATH as the default value.

* docs: update changelog